Separate openai_llm_cost.py
